requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
google-cloud-language==2.17.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
